+++ app/src/at-huybui/AndroidManifest.xml
@@ -19,8 +19,7 @@
             android:name="com.theartofdev.edmodo.cropper.CropImageActivity"
             android:theme="@style/Base.Theme.AppCompat" />
         <!-- optional (needed if default theme has no action bar) -->
-        <activity android:name=".w0.TestActivity"></activity>
-        <activity android:name=".w11.activity.ApiMainActivity" />
+        <activity android:name="com.asiantech.intern20summer1.w11.ui.activity.ApiMainActivity" />
         <activity android:name=".w6.secondactivity.ViewPagerTwoActivity" />
         <activity android:name=".w6.firstactivity.ViewPagerOneActivity" />
         <activity android:name=".w3.MainActivity" />
