+++ app/src/at-huybui/java/com/asiantech/intern20summer1/w11/data/api/apiservice/ApiPostService.kt
@@ -1,11 +1,12 @@
-package com.asiantech.intern20summer1.w11.api
+package com.asiantech.intern20summer1.w11.data.api.apiservice
 
-import com.asiantech.intern20summer1.w11.models.PostItem
-import com.asiantech.intern20summer1.w11.models.ResponseLike
-import com.asiantech.intern20summer1.w11.models.ResponsePost
+import com.asiantech.intern20summer1.w11.data.models.PostItem
+import com.asiantech.intern20summer1.w11.data.models.ResponseLike
+import com.asiantech.intern20summer1.w11.data.models.ResponsePost
 import io.reactivex.Observable
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
+import retrofit2.Response
 import retrofit2.http.*
 
 /**
@@ -30,7 +31,7 @@ interface ApiPostService {
         @Header("token") token: String,
         @Part image: MultipartBody.Part? = null,
         @Part("body") body: RequestBody
-    ): Observable<ResponsePost>
+    ): Observable<Response<ResponsePost>>
 
     @Multipart
     @POST(PART_UPDATE_POST)
@@ -39,20 +40,20 @@ interface ApiPostService {
         @Path("id") id: Int = 0,
         @Part image: MultipartBody.Part? = null,
         @Part("body") body: RequestBody
-    ): Observable<ResponsePost>
+    ): Observable<Response<ResponsePost>>
 
     @DELETE(PART_UPDATE_POST)
     fun deletePost(
         @Header("token") token: String,
         @Path("id") id: Int = 0
-    ): Observable<ResponsePost>
+    ): Observable<Response<ResponsePost>>
 
     @POST(PART_LIKE_POST)
     fun likePost(
         @Header("token") token: String,
         @Path("id") id: Int = 0
-    ): Observable<ResponseLike>
+    ): Observable<Response<ResponseLike>>
 
     @GET(PAST_GET_POSTS)
-    fun getPostLists(@Header("token") token: String): Observable<List<PostItem>>
+    fun getPostLists(@Header("token") token: String): Observable<Response<List<PostItem>>>
 }
