+++ app/src/at-huybui/java/com/asiantech/intern20summer1/w11/data/models/PostItem.kt
@@ -1,4 +1,4 @@
-package com.asiantech.intern20summer1.w11.models
+package com.asiantech.intern20summer1.w11.data.models
 
 import com.google.gson.annotations.SerializedName
