--- app/src/at-huybui/java/com/asiantech/intern20summer1/w11/models/SplashViewModel.kt
@@ -1,15 +1,0 @@
-package com.asiantech.intern20summer1.w11.models
-
-/**
- * Asian Tech Co., Ltd.
- * Intern20Summer1 Project.
- * Created by at-huybui on 14/09/2020.
- * This is SplashViewModel TODO("Not yet implemented").
- * It will TODO("Not yet implemented")
- */
-class SplashViewModel {
-
-
-//    private fun getAll
-}
-
