+++ app/src/at-huybui/java/com/asiantech/intern20summer1/w11/ui/fragment/PostDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.asiantech.intern20summer1.w11.fragment
+package com.asiantech.intern20summer1.w11.ui.fragment
 
 import android.Manifest
 import android.app.Activity
@@ -19,31 +19,30 @@ import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.DialogFragment
 import com.asiantech.intern20summer1.R
-import com.asiantech.intern20summer1.w11.activity.ApiMainActivity
-import com.asiantech.intern20summer1.w11.api.Api
-import com.asiantech.intern20summer1.w11.api.ApiPostService
-import com.asiantech.intern20summer1.w11.models.PostContent
-import com.asiantech.intern20summer1.w11.models.PostItem
-import com.asiantech.intern20summer1.w11.models.ResponsePost
+import com.asiantech.intern20summer1.w11.data.api.ApiClient
+import com.asiantech.intern20summer1.w11.data.models.PostContent
+import com.asiantech.intern20summer1.w11.data.repository.PostsRepository
+import com.asiantech.intern20summer1.w11.ui.activity.ApiMainActivity
+import com.asiantech.intern20summer1.w11.ui.viewmodel.HomeViewModel
 import com.asiantech.intern20summer1.w11.utils.AppUtils
 import com.asiantech.intern20summer1.w11.utils.FileInformation
-import com.bumptech.glide.Glide
 import com.google.gson.Gson
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.schedulers.Schedulers
 import kotlinx.android.synthetic.`at-huybui`.w10_dialog_fragment_post.*
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
 import okhttp3.RequestBody.Companion.asRequestBody
 import okhttp3.RequestBody.Companion.toRequestBody
-import retrofit2.Response
 
 /**
  * Asian Tech Co., Ltd.
  * Intern20Summer1 Project.
- * Created by at-huybui on 01/09/2020.
- * This is UpdateDialogFragment class. It is fragment to display update the post page
+ * Created by at-huybui on 02/09/2020.
+ * This is PostDialogFragment class. It is fragment to display add new post page
  */
 
-class UpdateDialogFragment : DialogFragment() {
+class PostDialogFragment : DialogFragment() {
 
     companion object {
         private const val REQUEST_IMAGE_CAPTURE = 100
@@ -51,28 +50,21 @@ class UpdateDialogFragment : DialogFragment() {
         private const val PERMISSION_REQUEST_CODE = 200
         private const val TYPE_IMAGE = "image/*"
         private const val TYPE_TEXT = "text"
-        internal fun newInstance(item: PostItem) = UpdateDialogFragment().apply {
-            postItem = item
-        }
+        internal fun newInstance() = PostDialogFragment()
     }
 
     internal var onPostClick: () -> Unit = {}
-    private lateinit var postItem: PostItem
-    private var callApi: ApiPostService? = null
     private var imageUri: Uri? = null
     private var isCameraAllowed = false
     private var isCheckGallery = false
+    private var viewModel: HomeViewModel? = null
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        callApi = Api.getInstance()?.create(ApiPostService::class.java)
-        dialog?.window?.apply {
-            setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
-            requestFeature(Window.FEATURE_NO_TITLE)
-        }
+        setupViewModel()
         return inflater.inflate(R.layout.w10_dialog_fragment_post, container, false)
     }
 
@@ -102,60 +94,43 @@ class UpdateDialogFragment : DialogFragment() {
     }
 
     private fun initView() {
-        initViewListener()
-        initDisPlayView()
+        initDialog()
+        initListener()
     }
 
 
-    private fun initViewListener() {
+    private fun initListener() {
         btnBack?.setOnClickListener {
             dialog?.dismiss()
         }
 
         btnPost?.setOnClickListener {
-            handleUpdateContent()
+            handlePostContent()
         }
         handleForAvatarImage()
     }
 
-    private fun initDisPlayView() {
-        if (postItem.image.isNotEmpty()) {
-            Glide.with(requireContext())
-                .load(Api.IMAGE_URL + postItem.image)
-                .into(imgContent)
-        }
-        tvTitle?.text = getString(R.string.w10_edit_post)
-        edtContent?.setText(postItem.content)
-    }
-
-    private fun handleUpdateContent() {
+    private fun handlePostContent() {
         progressBar?.visibility = View.VISIBLE
         val postJson = Gson().toJson(PostContent(edtContent?.text.toString())).toString()
         val body = postJson.toRequestBody(TYPE_TEXT.toMediaTypeOrNull())
         val token = AppUtils().getToken(requireContext())
-//        callApi?.updatePost(token, postItem.id, createMultiPartBody(), body)
-//            ?.enqueue(object : retrofit2.Callback<ResponsePost> {
-//                override fun onResponse(
-//                    call: retrofit2.Call<ResponsePost>,
-//                    response: Response<ResponsePost>
-//                ) {
-//
-//                    if (response.body()?.message == Api.MESSAGE_UPDATE_POST_SUCCESS) {
-//                        val text = getString(R.string.w10_update_complete)
-//                        ApiMainActivity().showToast(requireContext(), text)
-//                        onPostClick.invoke()
-//                        dialog?.dismiss()
-//                    } else {
-//                        val text = getString(R.string.w10_error_update)
-//                        ApiMainActivity().showToast(requireContext(), text)
-//                    }
-//                    progressBar?.visibility = View.INVISIBLE
-//                }
-//
-//                override fun onFailure(call: retrofit2.Call<ResponsePost>, t: Throwable) {
-//                    t.printStackTrace()
-//                }
-//            })
+        viewModel
+            ?.createPost(token, createMultiPartBody(), body)
+            ?.subscribeOn(Schedulers.io())
+            ?.observeOn(AndroidSchedulers.mainThread())
+            ?.subscribe { response ->
+                if (response.body()?.message == ApiClient.MESSAGE_CREATE_POST_SUCCESS) {
+                    val text = getString(R.string.w10_complete_post)
+                    ApiMainActivity().showToast(requireContext(), text)
+                    onPostClick.invoke()
+                    dialog?.dismiss()
+                } else {
+                    val text = getString(R.string.w10_error_post)
+                    ApiMainActivity().showToast(requireContext(), text)
+                }
+                progressBar?.visibility = View.INVISIBLE
+            }
     }
 
     private fun createMultiPartBody(): MultipartBody.Part? {
@@ -203,11 +178,10 @@ class UpdateDialogFragment : DialogFragment() {
      */
     private fun openCamera() {
         val values = ContentValues()
-        imageUri =
-            (activity as ApiMainActivity).contentResolver.insert(
-                MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
-                values
-            )
+        imageUri = (activity as ApiMainActivity).contentResolver.insert(
+            MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
+            values
+        )
         val cameraIntent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
         cameraIntent.putExtra(MediaStore.EXTRA_OUTPUT, imageUri)
         startActivityForResult(
@@ -221,12 +195,8 @@ class UpdateDialogFragment : DialogFragment() {
      */
     private fun openGallery() {
         val intentGallery = Intent(Intent.ACTION_PICK)
-        intentGallery.type =
-            TYPE_IMAGE
-        startActivityForResult(
-            intentGallery,
-            REQUEST_SELECT_IMAGE_IN_ALBUM
-        )
+        intentGallery.type = TYPE_IMAGE
+        startActivityForResult(intentGallery, REQUEST_SELECT_IMAGE_IN_ALBUM)
     }
 
     private fun isCheckCameraPermission(): Boolean {
@@ -276,5 +246,16 @@ class UpdateDialogFragment : DialogFragment() {
                 }
             }
         }
+    }
+
+    private fun initDialog() {
+        dialog?.window?.apply {
+            setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
+            requestFeature(Window.FEATURE_NO_TITLE)
+        }
+    }
+
+    private fun setupViewModel() {
+        viewModel = HomeViewModel(PostsRepository())
     }
 }
