+++ app/src/at-huybui/java/com/asiantech/intern20summer1/w11/ui/fragment/SignInFragment.kt
@@ -1,4 +1,4 @@
-package com.asiantech.intern20summer1.w11.fragment
+package com.asiantech.intern20summer1.w11.ui.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.Fragment
 import com.asiantech.intern20summer1.R
-import com.asiantech.intern20summer1.w11.activity.ApiMainActivity
-import com.asiantech.intern20summer1.w11.api.Api
-import com.asiantech.intern20summer1.w11.api.ApiAccountService
-import com.asiantech.intern20summer1.w11.api.ErrorUtils
-import com.asiantech.intern20summer1.w11.models.Account
+import com.asiantech.intern20summer1.w11.ui.activity.ApiMainActivity
+import com.asiantech.intern20summer1.w11.data.api.ApiClient
+import com.asiantech.intern20summer1.w11.data.api.apiservice.ApiAccountService
+import com.asiantech.intern20summer1.w11.data.api.ErrorUtils
+import com.asiantech.intern20summer1.w11.data.models.Account
 import com.asiantech.intern20summer1.w11.utils.AppUtils
 import kotlinx.android.synthetic.`at-huybui`.w10_fragment_sign_in.*
 import retrofit2.Response
@@ -36,7 +36,7 @@ class SignInFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        callApi = Api.getInstance()?.create(ApiAccountService::class.java)
+        callApi = ApiClient.getClientInstance()?.create(ApiAccountService::class.java)
         return inflater.inflate(R.layout.w10_fragment_sign_in, container, false)
     }
 
@@ -82,7 +82,7 @@ class SignInFragment : Fragment() {
                     }
                 } else {
                     val error = ErrorUtils().parseError(response)
-                    if (error?.message == Api.MESSAGE_LOGIN_INCORRECT) {
+                    if (error?.message == ApiClient.MESSAGE_LOGIN_INCORRECT) {
                         AppUtils().showToast(requireContext(),
                             getString(R.string.w10_email_or_pass_invalid))
                     }
