--- app/src/at-huybui/java/com/asiantech/intern20summer1/w11/utils/Resource.kt
@@ -1,27 +1,0 @@
-package com.asiantech.intern20summer1.w11.utils
-
-data class Resource<out T>(val status: Status, val data: T?, val message: String?) {
-
-    companion object {
-
-        fun <T> success(data: T?): Resource<T> {
-            return Resource(Status.SUCCESS, data, null)
-        }
-
-        fun <T> error(msg: String, data: T?): Resource<T> {
-            return Resource(Status.ERROR, data, msg)
-        }
-
-        fun <T> loading(data: T?): Resource<T> {
-            return Resource(Status.LOADING, data, null)
-        }
-
-    }
-
-}
-
-enum class Status {
-    SUCCESS,
-    ERROR,
-    LOADING
-}
