+++ app/src/at-huybui/res/layout/w10_activity_api.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.asiantech.intern20summer1.w11.activity.ApiMainActivity">
+    tools:context=".w11.ui.activity.ApiMainActivity">
 
     <FrameLayout
         android:id="@+id/containerMain"
